guards

+++ 3_remove_gp2.sh
@@ -1,2 +1,4 @@
 #!/bin/sh
-kubectl delete storageclass gp2
+if [ -n "$(kubectl get sc gp2 2>/dev/null)" ]; then
+	kubectl delete storageclass gp2
+fi
